+++ loky/reusable_executor.py
@@ -11,6 +11,7 @@ import multiprocessing as mp
 
 from .process_executor import ProcessPoolExecutor, EXTRA_QUEUED_CALLS
 from .process_executor import _SafeQueue
+from .backend.context import cpu_count
 from .backend.queues import SimpleQueue
 from .backend import get_context
 
@@ -168,6 +169,6 @@ class ReusablePoolExecutor(ProcessPoolExecutor):
     def _setup_queue(self, job_reducers, result_reducers):
         # As this executor can be resized, use a large queue size to avoid
         # underestimating capacity and introducing overhead
-        queue_size = 2 * mp.cpu_count() + EXTRA_QUEUED_CALLS
+        queue_size = 2 * cpu_count() + EXTRA_QUEUED_CALLS
         super(ReusablePoolExecutor, self)._setup_queue(
             job_reducers, result_reducers, queue_size=queue_size)